Rename namespace FlatbufferMetadata to AccessMessageMetadata
