+++ src/AccessMessageMetadata/ev42/ev42_Extractor.h
@@ -7,24 +7,20 @@
 //
 // Screaming Udder!                              https://esss.se
 
-/** Copyright (C) 2018 European Spallation Source ERIC */
-
-/// \file
-/// \brief Define classes required for chopper time stamp writing.
-
-#pragma once
-
 #include "FlatbufferReader.h"
 
-namespace FlatbufferMetadata {
+namespace AccessMessageMetadata {
 using FlatbufferMessage = FileWriter::FlatbufferMessage;
-using FBReaderBase = FileWriter::FlatbufferReader;
-
-/// See parent class for documentation.
-class tdct_Extractor : public FBReaderBase {
+class ev42_Extractor : public FileWriter::FlatbufferReader {
 public:
+  ev42_Extractor() = default;
+  ~ev42_Extractor() = default;
   bool verify(FlatbufferMessage const &Message) const override;
   std::string source_name(FlatbufferMessage const &Message) const override;
   uint64_t timestamp(FlatbufferMessage const &Message) const override;
+
+private:
+  SharedLogger Logger = spdlog::get("filewriterlogger");
 };
-} // namespace FlatbufferMetadata
+
+} // namespace AccessMessageMetadata
