+++ src/AccessMessageMetadata/f142/f142_Extractor.h
@@ -7,24 +7,17 @@
 //
 // Screaming Udder!                              https://esss.se
 
-/** Copyright (C) 2018 European Spallation Source ERIC */
-
-/// \file
-/// \brief Define classes required to implement the ADC file writing module.
-
 #pragma once
 
 #include "FlatbufferReader.h"
 
-namespace FlatbufferMetadata {
+namespace AccessMessageMetadata {
 using FlatbufferMessage = FileWriter::FlatbufferMessage;
-using FBReaderBase = FileWriter::FlatbufferReader;
 
-/// See parent class for documentation.
-class senv_Extractor : public FBReaderBase {
+class f142_Extractor : public FileWriter::FlatbufferReader {
 public:
   bool verify(FlatbufferMessage const &Message) const override;
   std::string source_name(FlatbufferMessage const &Message) const override;
   uint64_t timestamp(FlatbufferMessage const &Message) const override;
 };
-} // namespace FlatbufferMetadata
+} // namespace AccessMessageMetadata
