+++ src/tests/writer_module/ev42_WriterTests.cpp
@@ -16,7 +16,7 @@
 
 #include <utility>
 
-#include "fb_metadata_extractors/ev42/ev42_Extractor.h"
+#include "AccessMessageMetadata/ev42/ev42_Extractor.h"
 #include "helpers/HDFFileTestHelper.h"
 #include "helpers/SetExtractorModule.h"
 #include "writer_modules/ev42/ev42_Writer.h"
@@ -116,7 +116,7 @@ public:
   void SetUp() override {
     File = HDFFileTestHelper::createInMemoryTestFile("EventWriterTestFile.nxs");
     TestGroup = File.H5File.root().create_group(TestGroupName);
-    setExtractorModule<FlatbufferMetadata::ev42_Extractor>("ev42");
+    setExtractorModule<AccessMessageMetadata::ev42_Extractor>("ev42");
   };
 
   void TearDown() override { File.close(); };
